requirements.txt
# requirements.txt
# 翻訳ツール（PPTX / Excel / テキスト対応）用依存関係

python-pptx>=0.6.21
openpyxl>=3.0.10
deep-translator>=1.11.4
pillow>=9.0.0
requests>=2.28.0
